% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cells_format.R
\name{add_cells_format}
\alias{add_cells_format}
\title{Conditionally apply a function}
\usage{
add_cells_format(mmtable, ...)
}
\arguments{
\item{variable}{a column of the data frame from which header values will be constructed}
}
\value{
format list
}
\description{
Conditionally apply a function
}
