% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header_.R
\name{add_header_left}
\alias{add_header_left}
\title{Add a left header}
\usage{
add_header_left(mmtable, ...)
}
\arguments{
\item{mmtable}{an mmtable object}

\item{...}{arguments for header_left}
}
\value{
an mmtable
}
\description{
Add a left header
}
