% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header_.R
\name{add_header_top}
\alias{add_header_top}
\title{Add a top header}
\usage{
add_header_top(mmtable, ...)
}
\arguments{
\item{mmtable}{an mmtable object}

\item{...}{arguments for header_top}
}
\value{
an mmtable
}
\description{
Add a top header
}
