% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_functions.R
\name{add_spanner}
\alias{add_spanner}
\title{Add a spanner to a table}
\usage{
add_spanner(gm_table2, spanner_list)
}
\arguments{
\item{gm_table2}{a mmtable}

\item{spanner_list}{a list of spannner calls}
}
\value{
mmtable
}
\description{
Add a spanner to a table
}
