% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_format.R
\name{add_table_format}
\alias{add_table_format}
\title{Add formatting to a table}
\usage{
add_table_format(mmtable, ...)
}
\arguments{
\item{mmtable}{an mmtable object}

\item{...}{arguments for table_format}
}
\value{
an mmtable
}
\description{
Add formatting to a table
}
