% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_title.R
\name{add_table_source_note}
\alias{add_table_source_note}
\title{Conditionally apply a function}
\usage{
add_table_source_note(mmtable, ...)
}
\arguments{
\item{variable}{a column of the data frame from which header values will be constructed}
}
\value{
format list
}
\description{
Conditionally apply a function
}
