% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_format.R
\name{apply_format}
\alias{apply_format}
\title{Apply formating to a mmtable}
\usage{
apply_format(mmtable, format_list)
}
\arguments{
\item{mmtable}{an mmtable}

\item{format_list}{a list of formating commands.}
}
\value{
mmtable
}
\description{
Apply formating to a mmtable
}
