% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_formats.R
\name{apply_formats}
\alias{apply_formats}
\title{Apply list of formating commands to a mmtable}
\usage{
apply_formats(mmtable)
}
\arguments{
\item{mmtable}{an mmtable}
}
\value{
mmtable
}
\description{
Apply list of formating commands to a mmtable
}
