% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_if.R
\name{apply_if}
\alias{apply_if}
\title{Conditionally apply a function}
\usage{
apply_if(predicate, func, data, ...)
}
\arguments{
\item{predicate}{a predicate}

\item{func}{a function to apply}

\item{data}{data used by the function or otherwise returned}

\item{...}{other objects involved in evaluation of the function}
}
\value{
data
}
\description{
Conditionally apply a function
}
