% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cells_format.R
\name{cells_format}
\alias{cells_format}
\title{Conditionally apply a function}
\usage{
cells_format(cell_predicate, ...)
}
\arguments{
\item{cell_predicate}{an expression on the source data frame identifying which cells should be formatted.}

\item{...}{dots.}
}
\value{
format list
}
\description{
Conditionally apply a function
}
