% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_max_rects.R
\name{extract_max_rects}
\alias{extract_max_rects}
\title{Extract the largest binary rectary from a list of rows and cols}
\usage{
extract_max_rects(extraction_list)
}
\arguments{
\item{extraction_list}{a list containting a dataframe of rows and cols and a list of already extracted rows and cols.}
}
\value{
extraction_list
}
\description{
Extract the largest binary rectary from a list of rows and cols
}
