% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_a_loc.R
\name{format_a_loc}
\alias{format_a_loc}
\title{Conditionally apply a function}
\usage{
format_a_loc(mmtable, loc, format_list)
}
\arguments{
\item{mmtable}{an mmtable}

\item{loc}{a list containing a row and col items}

\item{format_list}{a list of formatting commands}
}
\value{
format list
}
\description{
Conditionally apply a function
}
