% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{gapminder_mm}
\alias{gapminder_mm}
\title{Gapminder data.}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
  \item{continent}{price, in US dollars}
  \item{country}{weight of the diamond, in carats}
  \item{value}{weight of the diamond, in carats}
  \item{var}{weight of the diamond, in carats}
  \item{year}{weight of the diamond, in carats}
  ...
}
continent country value var year
}
\source{
\url{http://www.diamondse.info/}
}
\usage{
gapminder_mm
}
\description{
Excerpt of the Gapminder data on life expectancy, GDP per capita, and population by country.
}
\keyword{datasets}
