% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_random_ids.R
\name{gen_random_ids}
\alias{gen_random_ids}
\title{Generate a random string}
\usage{
gen_random_ids(n, length)
}
\arguments{
\item{n}{number of ids}

\item{length}{length of id}
}
\value{
format list
}
\description{
Generate a random string
}
