% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_locations.R
\name{get_locations}
\alias{get_locations}
\title{Identify which locations are to be formatted.}
\usage{
get_locations(
  mmtable,
  header = NULL,
  func,
  cell_predicate = NULL,
  scope = "cell"
)
}
\arguments{
\item{mmtable}{an mmtable}

\item{header}{header from original data}

\item{func}{the original formatting function applied}

\item{cell_predicate}{predicate to determine locations}

\item{scope}{string determining how widely formating will apply. Options
include "cell", "headers" and "table"}
}
\value{
locations
}
\description{
Identify which locations are to be formatted.
}
