% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_max_area_bf.R
\name{get_max_area_bf}
\alias{get_max_area_bf}
\title{An algorthm to identify the largest rectangle in a histogram}
\usage{
get_max_area_bf(heights)
}
\arguments{
\item{heights}{a list of the heights of the histogram.}
}
\value{
list
}
\description{
An algorthm to identify the largest rectangle in a histogram
}
