% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_row_header_names.R
\name{get_row_header_names}
\alias{get_row_header_names}
\title{Get row header names.}
\usage{
get_row_header_names(df, col_header_df_01)
}
\arguments{
\item{df}{a dataframe}

\item{col_header_df_01}{a dataframe of col headers}
}
\value{
a vector of names
}
\description{
Get row header names.
}
