% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header_format.R
\name{header_format}
\alias{header_format}
\title{Conditionally apply a function}
\usage{
header_format(header, ...)
}
\arguments{
\item{header}{the nme of a variable that is currently a header in the table}

\item{...}{formating instructions}
}
\value{
format list
}
\description{
Conditionally apply a function
}
