% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header_.R
\name{header_left}
\alias{header_left}
\title{Conditionally apply a function}
\usage{
header_left(variable)
}
\arguments{
\item{variable}{a column of the data frame from which header values will be constructed}
}
\value{
format list
}
\description{
Conditionally apply a function
}
