% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header_merged_cols.R
\name{header_merged_cols}
\alias{header_merged_cols}
\title{Merge header columns}
\usage{
header_merged_cols(...)
}
\arguments{
\item{...}{formating instructions}
}
\value{
format list
}
\description{
Merge header columns
}
