% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_mmtable.R
\name{knit_print.mmtable}
\alias{knit_print.mmtable}
\title{Print an mmtable object}
\usage{
knit_print.mmtable(x, ...)
}
\arguments{
\item{x}{an mmtable}

\item{...}{other things}
}
\description{
Print an mmtable object
}
