% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmtable.R
\name{mmtable}
\alias{mmtable}
\title{Conditionally apply a function}
\usage{
mmtable(data, table_data, table_name = NULL)
}
\arguments{
\item{data}{a data frame}

\item{table_data}{the name of the column whose values will form the data cells.}

\item{table_name}{the name of the table - used for labelling when joined with another table - randomly generated if not provided.}
}
\value{
mmtable
}
\description{
Conditionally apply a function
}
