% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapminder_funcs.R
\name{model_lexp}
\alias{model_lexp}
\title{Conditionally apply a function}
\usage{
model_lexp(data, country_input)
}
\arguments{
\item{data}{an mmtable}

\item{country_input}{a list containing a row and col items}
}
\description{
Conditionally apply a function
}
