% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus_mmtable_.R
\name{+.mmtable}
\alias{+.mmtable}
\title{Conditionally apply a function}
\usage{
\method{generic}{class}(mmtable1, mmtable2)
}
\arguments{
\item{mmtable1}{an mmtable}

\item{mmtable2}{an mmtable}
}
\value{
an mmtable
}
\description{
Conditionally apply a function
}
