% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_mmtable.R
\name{print.mmtable}
\alias{print.mmtable}
\title{Print an mmtable object}
\usage{
\method{print}{mmtable}(x, ...)
}
\arguments{
\item{x}{an mmtable}

\item{...}{other things}
}
\description{
Print an mmtable object
}
