% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_functions.R
\name{set_class}
\alias{set_class}
\title{Set the class of an object}
\usage{
set_class(object, class)
}
\arguments{
\item{object}{a object}

\item{class}{new class of object}
}
\value{
mmtable
}
\description{
Set the class of an object
}
