% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_cells_to_vector.R
\name{single_cells_to_vector}
\alias{single_cells_to_vector}
\title{Convert a list of cells to a list of cell ranges}
\usage{
single_cells_to_vector(row_col_list)
}
\arguments{
\item{row_col_list}{list of cells}
}
\description{
Convert a list of cells to a list of cell ranges
}
