% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_and_select_columns.R
\name{sort_and_select_columns}
\alias{sort_and_select_columns}
\title{Sort and select columns}
\usage{
sort_and_select_columns(data, ...)
}
\arguments{
\item{data}{a data frame}

\item{...}{list of data to load into envrionment}
}
\value{
mmtable
}
\description{
Sort and select columns
}
