% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_functions.R
\name{spannerize}
\alias{spannerize}
\title{Convert header to spanner}
\usage{
spannerize(gm_table2, n)
}
\arguments{
\item{gm_table2}{a mmtable}

\item{n}{number of spanners}
}
\value{
mmtable
}
\description{
Convert header to spanner
}
