% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread_col_headers.R
\name{spread_col_headers}
\alias{spread_col_headers}
\title{Spread rows to columns}
\usage{
spread_col_headers(data, ...)
}
\arguments{
\item{data}{a data frame}

\item{...}{list of data to load into envrionment}
}
\value{
a data frame
}
\description{
Spread rows to columns
}
