% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{student_df}
\alias{student_df}
\title{A minimal data set of student subjects and grades}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
  \item{subject}{}
  \item{class}{}
  \item{student}{}
  \item{value}{}
  \item{grade}{}
  ...
}
}
\usage{
student_df
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
