% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{student_df_2}
\alias{student_df_2}
\title{Prices of 50,000 round cut diamonds.}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
  \item{class}{price, in US dollars}
  \item{grade}{price, in US dollars}
  \item{student}{price, in US dollars}
  \item{subject}{price, in US dollars}
  \item{value}{price, in US dollars}
  ...
}
}
\source{
\url{http://www.diamondse.info/}
}
\usage{
student_df_2
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
