% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_spanner.R
\name{style_spanner}
\alias{style_spanner}
\title{Style a spanner}
\usage{
style_spanner(table, format)
}
\arguments{
\item{table}{an gt_tbl}

\item{format}{formatting}
}
\value{
a data frame
}
\description{
Style a spanner
}
