% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_format.R
\name{table_format}
\alias{table_format}
\title{Conditionally apply a function}
\usage{
table_format(locations, ...)
}
\arguments{
\item{locations}{indicating which cells should be fomrated}

\item{...}{formating instructions}
}
\value{
format list
}
\description{
Conditionally apply a function
}
