% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapminder_funcs.R
\name{table_model_wide}
\alias{table_model_wide}
\title{Conditionally apply a function}
\usage{
table_model_wide(model_df)
}
\arguments{
\item{model_df}{an mmtable}
}
\description{
Conditionally apply a function
}
