% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_title.R
\name{table_source_note}
\alias{table_source_note}
\title{Table source note}
\usage{
table_source_note(...)
}
\arguments{
\item{...}{arguments of gt::tab_source_note}
}
\value{
format list
}
\description{
Table source note
}
