% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_title.R
\name{table_title}
\alias{table_title}
\title{Conditionally apply a function}
\usage{
table_title(...)
}
\arguments{
\item{...}{formating instructions}
}
\value{
format list
}
\description{
Conditionally apply a function
}
