% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_spanner.R
\name{transfer_spanner}
\alias{transfer_spanner}
\title{Transfer spanner from one table to another}
\usage{
transfer_spanner(gt_01, gt_02)
}
\arguments{
\item{mmtable}{an mmtable object}

\item{...}{arguments of gt::tab_source_note}
}
\value{
an mmtable
}
\description{
Transfer spanner from one table to another
}
