% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cells_format.R
\name{add_cells_format}
\alias{add_cells_format}
\title{Add cells formatting}
\usage{
add_cells_format(mmtable, ...)
}
\arguments{
\item{mmtable}{an mmtable object}

\item{...}{arguments for add_cells_format}
}
\value{
an mmtable
}
\description{
Add cells formatting
}
