% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header_format.R
\name{add_header_format}
\alias{add_header_format}
\title{Add formatting to a header}
\usage{
add_header_format(mmtable, ...)
}
\arguments{
\item{mmtable}{an mmtable object}

\item{...}{arguments for add_header_format}
}
\value{
an mmtable
}
\description{
Add formatting to a header
}
