% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header_.R
\name{add_header_left_top}
\alias{add_header_left_top}
\title{Add a left top header}
\usage{
add_header_left_top(mmtable, ...)
}
\arguments{
\item{mmtable}{an mmtable object}

\item{...}{arguments for header_left_top}
}
\value{
an mmtable object
}
\description{
Add a left top header
}
\examples{
\dontrun{

library(tidyverse)

gm_table_piped <-
  gapminder_mm \%>\%
  filter(var != "Life expectancy") \%>\%
  mmtable(table_data = value, use_defaul_formats = T) \%>\%
  add_header_top(year) \%>\%
  add_header_left(country) \%>\%
  add_header_top_left(var)  \%>\%
  add_header_left_top(continent)  \%>\%
  add_cells_format(cell_predicate = T, style = list(cell_text(align = "right"))) \%>\%
  add_header_format(header = year, style = list(cell_text(align = "right"))) \%>\%
  add_header_format("all_cols", style = list(cell_text(weight = "bolder"))) \%>\%
  add_header_format("all_rows", style = list(cell_text(weight = "bolder"))) \%>\%
  add_table_format(
    locations = cells_body(rows = c(1,3,5,7,9,11)),
    style = list(cell_borders(sides = "top",color = "grey"))) \%>\%
  add_table_source_note(source_note = "Excerpt of the Gapminder dataset." )

gm_table_piped
}
}
