% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header_.R
\name{add_header_top}
\alias{add_header_top}
\title{Conditionally apply a function}
\usage{
add_header_top(mmtable, ...)
}
\arguments{
\item{variable}{a column of the data frame from which header values will be constructed}
}
\value{
format list
}
\description{
Conditionally apply a function
}
