% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_title.R
\name{add_table_source_note}
\alias{add_table_source_note}
\title{Add a table source note}
\usage{
add_table_source_note(mmtable, ...)
}
\arguments{
\item{mmtable}{an mmtable object}

\item{...}{arguments for table_source_note}
}
\value{
an mmtable
}
\description{
Add a table source note
}
