% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cells_format.R
\name{cells_format}
\alias{cells_format}
\title{Conditionally apply a function}
\usage{
cells_format(cell_predicate, ...)
}
\arguments{
\item{cell_predicate}{an expression on the source data frame identifying which cells should be formatted.}

\item{...}{dots.}
}
\value{
format list
}
\description{
Conditionally apply a function
}
\examples{
\dontrun{
gm_table_formatted <-
  gapminder_mm \%>\%
  filter(var != "Life expectancy") \%>\%
  mmtable(table_data = value) +
  header_top(year) +
  header_left(country) +
  header_top_left(var)  +
  header_left_top(continent)  +
  cells_format(cell_predicate = T, style = list(cell_text(align = "right"))) +
  header_format(header = year, style = list(cell_text(align = "right"))) +
  header_format("all_cols", style = list(cell_text(weight = "bolder"))) +
  header_format("all_rows", style = list(cell_text(weight = "bolder"))) +
  table_format(
    locations = cells_body(rows = c(1,3,5,7,9,11)),
    style = list(cell_borders(sides = "top",color = "grey"))) +
  table_source_note(source_note = "Excerpt of the Gapminder data" )
}
}
