% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagonalize.R
\name{diagonalize}
\alias{diagonalize}
\title{Conditionally apply a function}
\usage{
diagonalize(df, var_char)
}
\arguments{
\item{df}{a data frame}

\item{var_char}{the column to be diagonalized}
}
\value{
data frame
}
\description{
Conditionally apply a function
}
