% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_functions.R
\name{get_spanner_html_text}
\alias{get_spanner_html_text}
\title{Get the html associated with a spanner}
\usage{
get_spanner_html_text(table)
}
\arguments{
\item{table}{an mmtable}
}
\value{
html
}
\description{
Get the html associated with a spanner
}
