% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header_format.R
\name{header_format}
\alias{header_format}
\title{Format the cells of a header.}
\usage{
header_format(header, ...)
}
\arguments{
\item{header}{the name of a variable that is currently a header in the table}

\item{...}{a list of styles, with format required by gt::tab_style style argument. https://gt.rstudio.com/reference/tab_style.html}
}
\value{
format list
}
\description{
Format the cells of a header.
}
