% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header_.R
\name{header_top}
\alias{header_top}
\title{Add a top header}
\usage{
header_top(variable)
}
\arguments{
\item{variable}{a column of the data frame from which header values will be constructed}
}
\value{
an mmtable object
}
\description{
Add a top header
}
\examples{
\dontrun{

library(tidyverse)

row_list <- cells_body(rows = c(1,3,5,7,9,11))
style_list <- list(cell_borders(sides = "top",color = "grey"))
gm_df <- gapminder_mm \%>\% dplyr::filter(var != "Life expectancy")

gm_table <-
  gm_df \%>\%
  mmtable(table_data = value) +
  header_top(year) +
  header_left(country) +
  header_top_left(var)  +
  header_left_top(continent)  +
  table_format(
    locations = row_list,
    style = style_list)
}
}
