% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmtable.R
\name{mmtable}
\alias{mmtable}
\title{Conditionally apply a function}
\usage{
mmtable(data, table_data, table_name = NULL, use_defaul_formats = TRUE)
}
\arguments{
\item{data}{a data frame}

\item{table_data}{the name of the column whose values will form the data cells.}

\item{table_name}{the name of the table - used for labelling when joined with another table - randomly generated if not provided.}
}
\value{
mmtable
}
\description{
Conditionally apply a function
}
\examples{
\dontrun{

library(tidyverse)

row_list <- cells_body(rows = c(1,3,5,7,9,11))
style_list <- list(cell_borders(sides = "top",color = "grey"))
gm_df <- gapminder_mm \%>\% filter(var != "Life expectancy")

gm_table <-
  gm_df \%>\%
  mmtable(table_data = value) +
  header_top(year) +
  header_left(country) +
  header_top_left(var)  +
  header_left_top(continent)  +
  table_format(
    locations = row_list,
    style = style_list)
}
}
