% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_mmtable.R
\name{print.mmtable}
\alias{print.mmtable}
\title{Conditionally apply a function}
\usage{
\method{generic}{class}(mmtable)
}
\arguments{
\item{e1}{an mmtable}

\item{e2}{an mmtable}
}
\value{
an mmtable
}
\description{
Conditionally apply a function
}
