% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide_mmtable.R
\name{/.mmtable}
\alias{/.mmtable}
\title{Concatenate tables horizontally}
\usage{
\method{/}{mmtable}(e1, e2)
}
\arguments{
\item{e1}{an mmtable}

\item{e2}{an mmtable}
}
\value{
an mmtable
}
\description{
Concatenate tables horizontally
}
