% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_first_col_header.R
\name{style_first_col_header}
\alias{style_first_col_header}
\title{Style the first col header}
\usage{
style_first_col_header(table, format)
}
\arguments{
\item{data}{a data frame}

\item{...}{list of data to load into envrionment}
}
\value{
a data frame
}
\description{
Style the first col header
}
