% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_format.R
\name{table_format}
\alias{table_format}
\title{Format a mmtable object.}
\usage{
table_format(locations, ...)
}
\arguments{
\item{locations}{indicating which cells should be formatted.}

\item{...}{formating instructions}
}
\value{
format list
}
\description{
Format a mmtable object.
}
\examples{
\dontrun{

library(tidyverse)

row_list <- cells_body(rows = c(1,3,5,7,9,11))
style_list <- list(cell_borders(sides = "top",color = "grey"))
gm_df <- gapminder_mm \%>\% dplyr::filter(var != "Life expectancy")

gm_table <-
  gm_df \%>\%
  mmtable(table_data = value) +
  header_top(year) +
  header_left(country) +
  header_top_left(var)  +
  header_left_top(continent)  +
  table_format(
    locations = row_list,
    style = style_list)
}
}
