% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_title.R
\name{table_source_note}
\alias{table_source_note}
\title{Conditionally apply a function}
\usage{
table_source_note(...)
}
\arguments{
\item{...}{formating instructions}
}
\value{
format list
}
\description{
Conditionally apply a function
}
