% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiply_mmtable_.R
\name{*.mmtable}
\alias{*.mmtable}
\title{Integrate two tables horizontally}
\usage{
\method{*}{mmtable}(mmtable1, mmtable2)
}
\arguments{
\item{mmtable1}{an mmtable}

\item{mmtable2}{an mmtable}
}
\value{
an mmtable
}
\description{
Integrate two tables horizontally
}
